Add files without leakage
I went back to re-assess model performance and realized my technique for scaling/vectorizing used the whole dataset instead of just basing the behavior on the training set. It turns out this marginally improved performance for the attribute model and brought lyric accuracy above 70%. The next step here would be to recollect lyric data and try 8 classes with 1000 songs each or so.